Eliminate font/color duplication between renderers
Move shared font and color utilities to amux-term as the single source
of truth, replacing duplicate implementations in amux-render-soft and
amux-render-gpu:

- Add create_font_system() and measure_cell_width() to amux-term::font
- Add srgba_to_f32() to amux-term::color
- Delete resolve_color() copies from both renderers (import from amux-term)
- Delete srgba_to_rgba8() copy from soft renderer (import from amux-term)
- Delete measure_cell_width() from both renderers (import from amux-term)
- Delete font init boilerplate from both renderers (use create_font_system)

Net: -111 lines, 7 files changed.

Closes #44
